requirements.txt
kafka-python==2.0.2
kafka-python-ng==2.2.3
psycopg2==2.9.10
python-dotenv==1.0.1
flask=3.1.0
requests==2.32.3
